+++ icarus2-query-api/src/main/java/de/ims/icarus2/query/api/eval/UnaryOperations.java
@@ -74,7 +74,7 @@ public class UnaryOperations {
 		}
 
 		@Override
-		public TypeInfo getResultType() { return TypeInfo.LONG; }
+		public TypeInfo getResultType() { return TypeInfo.INTEGER; }
 
 		@Override
 		public Primitive<? extends Number> compute() {
@@ -106,7 +106,7 @@ public class UnaryOperations {
 		}
 
 		@Override
-		public TypeInfo getResultType() { return TypeInfo.LONG; }
+		public TypeInfo getResultType() { return TypeInfo.INTEGER; }
 
 		@Override
 		public Primitive<? extends Number> compute() {
